requirements.txt
Flask==2.3.3
torch==2.2.2
torchaudio==2.2.2
nemo_toolkit[asr]==1.23.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.4
setproctitle
